added theory in line

+++ TestKalkulatora/TestKalkulatora/Program.cs
@@ -11,7 +11,6 @@ namespace TestKalkulatora
 
         static void Main(string[] args)
         {
-
             Console.WriteLine("Press: 1 - dodawanie, 2 -odejmowanie, 3 - mnozenie,  4 - dzielenie" );
             var operationType = int.Parse(Console.ReadLine());
 
@@ -26,27 +25,22 @@ namespace TestKalkulatora
             {
                 result = calc.Add(x, y);
             }
-
-
+            
             if (operationType == 2)
             {
                 result = calc.Subtract(x, y);
             }
-
-
+            
             if (operationType == 3)
             {
                  result = calc.Multiply(x, y);
             }
-
-
+            
             if (operationType == 4)
             {
                  result = calc.Divide(x, y);
             }
-
-            
-
+                        
             Console.WriteLine("-------");
             Console.WriteLine(result);
             Console.ReadLine();
@@ -60,22 +54,22 @@ namespace TestKalkulatora
         {
         }
 
-        public double Add(int x, int y)
+        public double Add(double x, double y)
         {
             return x + y;
         }
 
-        public double Subtract(int x, int y)
+        public double Subtract(double x, double y)
         {
             return x - y;
         }
 
-        public double Multiply(int x, int y)
+        public double Multiply(double x, double y)
         {
             return x * y;
         }
 
-        public double Divide(int x, int y)
+        public double Divide(double x, double y)
         {
             return x / y;
         }
